contract: Divide file for maintainability
